Multiple changes
- Added a new angle named constructor
- Added polygon rasterization mode enumeration
- Added wrappers for PolygonMode, PolygonOffset, LineWidth, PointSize to context
- Added a new example showing Cel-shading
- Minor fixes in the drawing instructions initializers of Sphere and Torus

+++ include/oglplus/shapes/torus.hpp
@@ -171,8 +171,8 @@ public:
 				{
 					DrawOperation::Method::DrawElements,
 					PrimitiveType::TriangleStrip,
-					r * (_sections + 1) * 2,
-					(_sections + 1) * 2
+					GLuint(r * (_sections + 1) * 2),
+					GLuint((_sections + 1) * 2)
 				}
 			);
 		}
